assignment-42.2

+++ app/controllers/comments_controller.rb
@@ -7,35 +7,51 @@ class CommentsController < ApplicationController
   def create
     if params[:topic].present?
       @topic = Topic.find(params[:topic_id])
-    else
-      @post = Post.find(params[:post_id])
-    end
-    comment = @post.comments.new(comment_params)
-    comment.user = current_user
- 
-    if comment.save
-      flash[:notice] = "Comment saved successfully."
-      redirect_to [@post.topic, @post]
-    else
-      flash[:alert] = "Comment failed to save."
-      redirect_to [@post.topic, @post]
+      comment = @topic.comments.new(comment_params)
+      comment.user = current_user
+      if comment.save
+        flash[:notice] = "Comment saved successfully."
+        redirect_to [@topic]
+      else
+        flash[:alert] = "Comment failed to save."
+        redirect_to [@topic]
+      end
+    elsif params[:post].present?
+       @post = Post.find(params[:post_id])
+       comment = @post.comments.new(comment_params)
+       comment.user = current_user
+      if comment.save
+        flash[:notice] = "Comment saved successfully."
+        redirect_to [@post.topic, @post]
+      else
+        flash[:alert] = "Comment failed to save."
+        redirect_to [@post.topic, @post]
+      end
     end
   end
   
   def destroy
     if params[:topic].present?
       @topic = Topic.find(params[:topic_id])
-    else
+      comment = @topic.comments.find(params[:id])
+      if comment.destroy
+        flash[:notice] = "Comment was deleted successfully."
+        redirect_to [@topic]
+      else
+        flash[:alert] = "Comment couldn't be deleted. Try again."
+        redirect_to [@topic]
+      end
+      
+    elsif params[:post].present?
       @post = Post.find(params[:post_id])
-    end
-    comment = @post.comments.find(params[:id])
- 
-    if comment.destroy
-      flash[:notice] = "Comment was deleted successfully."
-      redirect_to [@post.topic, @post]
-    else
-      flash[:alert] = "Comment couldn't be deleted. Try again."
-      redirect_to [@post.topic, @post]
+      comment = @post.comments.find(params[:id])
+      if comment.destroy
+        flash[:notice] = "Comment was deleted successfully."
+        redirect_to [@post.topic, @post]
+      else
+        flash[:alert] = "Comment couldn't be deleted. Try again."
+        redirect_to [@post.topic, @post]
+      end
     end
   end
  
@@ -46,11 +62,6 @@ class CommentsController < ApplicationController
   end
   
   def authorize_user
-    if params[:topic].present?
-      @topic = Topic.find(params[:topic_id])
-    else
-      @post = Post.find(params[:post_id])
-    end
     comment = Comment.find(params[:id])
     unless current_user == comment.user || current_user.admin?
       flash[:alert] = "You do not have permission to delete a comment."
